requirements.txt
ttkbootstrap
Pillow
customtkinter